Implement structure TAF
