+++ babel.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-    presets: [
-        ['@babel/preset-env', {
-            targets: {
-                node: 12
-            }
-        }]
-    ]
-}
+  presets: [
+    ['@babel/preset-env', {
+      targets: {
+        node: 12,
+      },
+    }],
+  ],
+};
