+++ test/utils/pageobjects/page.js
@@ -3,14 +3,12 @@
 * that is shared across all page objects
 */
 export default class Page {
-    constructor() {
+  constructor() {
 
-    }
-    /**
+  }
+
+  /**
     * Opens a sub page of the page
     * @param path path of the sub page (e.g. /path/to/page.html)
     */
-    open (path) {
-        return browser.url(`https://the-internet.herokuapp.com/${path}`)
-    }
 }
